backend/requirements.txt
Werkzeug==1.0.1
Flask==1.1.2
pandas==1.2.3
pymongo==3.11.3
PyJWT==2.0.1
python-dotenv==0.17.0
flask_cors
pytz